# internal/handler/swagger/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "四川华电泸定电站 API",
    "description": "四川华电泸定电站后端服务 swagger 文档",
    "version": "v0.0.1"
  },
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/job/download/jobs": {
      "post": {
        "summary": "作业文件下载",
        "operationId": "DownloadJobs",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/byte"
              }
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "description": " 下载作业请求",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/DownloadJobsReq"
            }
          }
        ],
        "tags": [
          "job"
        ]
      }
    },
...
  },
  "definitions": {
...
    "DownloadJobsReq": {
      "type": "object",
      "properties": {
        "jobNumbers": {
          "type": "array",
          "items": {
            "type": "string"
          },
          "description": " 要下载的作业号列表"
        }
      },
      "title": "DownloadJobsReq",
      "required": [
        "jobNumbers"
      ]
    },
...
  },
  "securityDefinitions": {
    "apiKey": {
      "type": "apiKey",
      "description": "Enter JWT Bearer token **_only_**",
      "name": "Authorization",
      "in": "header"
    }
  }
}